web/requirements.txt
Django==2.1.7
gunicorn==19.9.0
psycopg2-binary==2.8.6
redis==3.2.1
PyJWT
